Move files to new folder
- Comment out generator attribute since the derivative generator is still an experimental feature and we do not want to generate unnecessary files.
